requirements.txt
python-docx
docxtpl
openpyxl
pandas
python-dotenv
azure-cognitiveservices-vision-customvision
